+++ src/questionnaire/questions.js
@@ -1,26 +1,38 @@
 const inquirer = require('inquirer');
 const fs = require('fs');
 
-const settings = require('../settings');
 const utils = require('../utils');
 
 /**
  * Run the prompts to get the details for the project
+ * @method promptProjectDetails
  * @param  {Object} defaults
  * @param  {String} defaults.projectName    The default name for the project
  * @param  {String} defaults.version        The default version for the project
  * @param  {String} defaults.license        The default license for the project
  * @param  {String} defaults.gitUserName    The git username setup for the project
  * @param  {String} defaults.gitUserEmail   The git username setup for the project
+ * @param  {String} defaults.template       The name for the default template
+ * @param  {Array} licenses                 The list of options for licenses
+ * @param  {Array} testingPkgs              The list of options for testingPkgs
+ * @param  {Array} templates                The list of options for templates
  * @return {Promise}
  */
-async function getProjectDetails(defaults) {
+async function promptProjectDetails(defaults, licenses, testingPkgs, templates) {
   return inquirer.prompt([
     {
       type: 'input',
       name: 'name',
       message: 'What is your project name?',
       default: defaults.projectName,
+    },
+
+    {
+      type: 'list',
+      name: 'template',
+      message: 'What kind of project are you creating?',
+      choices: templates,
+      default: defaults.template,
     },
 
     {
@@ -47,7 +59,7 @@ async function getProjectDetails(defaults) {
       type: 'list',
       name: 'license',
       message: 'Please select a license',
-      choices: settings.licenses,
+      choices: licenses,
       default: defaults.license,
     },
 
@@ -88,7 +100,7 @@ async function getProjectDetails(defaults) {
       type: 'checkbox',
       name: 'testPackages',
       message: 'Which test packages do you want to include?',
-      choices: settings.testingPkgs,
+      choices: testingPkgs,
     },
 
     {
@@ -101,9 +113,10 @@ async function getProjectDetails(defaults) {
 
 /**
  * Run the prompts to get the details for the remote git
+ * @method promptGitRemoteDetails
  * @return {Promise}
  */
-async function getGitRemoteDetails() {
+async function promptGitRemoteDetails() {
   return inquirer.prompt([
     {
       type: 'input',
@@ -121,21 +134,23 @@ async function getGitRemoteDetails() {
 
 /**
  * Run the prompts to geth the path for the auth file
+ * @method promptSettingsFile
+ * @param  {String} settingsPath  The default path for the settings file
  * @return {Promise}
  */
-async function getAuthFile() {
+async function promptSettingsFile(settingsPath) {
   return inquirer.prompt([
     {
       type: 'input',
-      name: 'authPath',
-      message: 'What is the path for the create-nodejs-project.json file?',
-      default: settings.authPath,
+      name: 'settingsPath',
+      message: 'What is the path for the create-nodejs-settings.json file?',
+      default: settingsPath,
       validate: (ans) => {
         const path = utils.files.resolvePath(ans);
         if (path && fs.existsSync(path)) {
           return true;
         }
-        return 'You should introduce a real path for the create-nodejs-project.json';
+        return 'You should introduce a real path for the create-nodejs-settings.json';
       },
     },
   ]);
@@ -143,10 +158,11 @@ async function getAuthFile() {
 
 /**
  * Run the prompts to get the github user
+ * @method promptGithubUser
  * @param  {String} user The current user on the auth file
  * @return {Promise}
  */
-async function getGithubUser(user) {
+async function promptGithubUser(user) {
   return inquirer.prompt([
     {
       type: 'input',
@@ -159,11 +175,12 @@ async function getGithubUser(user) {
 
 /**
  * Run the prompts to get the github token
+ * @method promptAuthToken
  * @param  {String} user  The current github user on the auth file
  * @param  {String} token The current github user on the auth file
  * @return {Promise}
  */
-async function getAuthToken(user, token) {
+async function promptAuthToken(user, token) {
   return inquirer.prompt([
     {
       type: 'input',
@@ -176,23 +193,28 @@ async function getAuthToken(user, token) {
 
 /**
  * Run the prompt to confirm if the user wants to update the token
+ * @method promptUpdateToken
  * @return {Promise}
  */
-async function confirmUpdateToken() {
+async function promptUpdateToken() {
   return inquirer.prompt([
     {
       type: 'confirm',
       name: 'updateToken',
-      message: 'Do you want to update the create-nodejs-project.json file with this token?',
+      message: 'Do you want to update the settings file with this token?',
     },
   ]);
 }
 
+/**
+ * The questions for the questionnaire
+ * @module questions
+ */
 module.exports = {
-  getProjectDetails,
-  getGitRemoteDetails,
-  getAuthFile,
-  getAuthToken,
-  confirmUpdateToken,
-  getGithubUser,
+  promptProjectDetails,
+  promptGitRemoteDetails,
+  promptSettingsFile,
+  promptAuthToken,
+  promptUpdateToken,
+  promptGithubUser,
 };