+++ src/template/index.js
@@ -3,14 +3,13 @@ const path = require('path');
 
 const utils = require('../utils');
 
-const TEMPLATE_PATH = path.join(__dirname, '..', '..', 'template');
-const LICENSES_PATH = path.join(__dirname, '..', '..', 'licenses');
-
 /**
- * Update a file using a dictionary
- * @param  {Object} dictionary    A key-value dictionary
- * @param  {String} [filePath=''] The path for the file
- */
+* Update a file using a dictionary
+* @method updateFile
+* @param  {Object} dictionary    A key-value dictionary
+* @param  {String} [filePath=''] The path for the file
+* @return {Promise}
+*/
 async function updateFile(dictionary, filePath = '') {
   const resolvedFilePath = utils.files.resolvePath(filePath);
 
@@ -23,17 +22,29 @@ async function updateFile(dictionary, filePath = '') {
 }
 
 /**
- * Copy the template folder recursively
- * @param  {String} templatePath The template path
- * @param  {String} destPath     The destination
- */
-async function copyTemplate(destPath) {
-  await utils.files.copyDirRecursive(TEMPLATE_PATH, destPath);
+* Copy the template folder recursively
+* @method copyTemplate
+* @param  {String} templatePath The template path
+* @param  {String} destPath     The destination
+* @return {Promise}
+*/
+async function copyTemplate(templatePath, destPath) {
+  await utils.files.copyDirRecursive(templatePath, destPath);
 }
 
-async function copyLicense(license, dictionary, destPath) {
+/**
+ * Read the license file, update its content with the project data and save it
+ * on the new destination
+ * @method copyLicense
+ * @param  {String} license    The license name
+ * @param  {Object} dictionary The dictionary used to update the license content
+ * @param  {String} originPath The path for the licenses folder
+ * @param  {String} destPath   The destination path
+ * @return {Promise}
+ */
+async function copyLicense(license, dictionary, originPath, destPath) {
   // Get the license file
-  const resolvedFilePath = path.join(LICENSES_PATH, license.replace(' ', '-'));
+  const resolvedFilePath = path.join(originPath, license.replace(' ', '-'));
   const originalFile = await fs.readFile(resolvedFilePath, 'utf8');
 
   // Replace with dictionary
@@ -45,6 +56,14 @@ async function copyLicense(license, dictionary, destPath) {
   console.log(`File ${licensePath} created`);
 }
 
+/**
+ * Update some of the template files with the project data
+ * @method updateTemplateFiles
+ * @param  {Object} dictionary The dictionary with the data to be updated on the
+ * template files
+ * @param  {String} destPath   The destination path
+ * @return {Promise}
+ */
 function updateTemplateFiles(dictionary, destPath) {
   const readmePath = path.join(destPath, 'README.md');
   const packagePath = path.join(destPath, 'package.json');
@@ -55,6 +74,10 @@ function updateTemplateFiles(dictionary, destPath) {
   ]);
 }
 
+/**
+ * The template handler
+ * @module template
+ */
 module.exports = {
   updateFile,
   copyTemplate,