+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.0
+* Settings is a global object handle from an accessible json file.
+* Includes some unit tests
+* Added CI with travis
+
 ## 0.2.0
 * Generate license file